test: move tests to test folder, so they are not added to dist
